refactor(infrastructure): finalize the casting layout and add the example
