+++ docs/examples/aws/kubernetes/terraform/casting.yaml
@@ -1,0 +1,13 @@
+apiVersion: v1alpha1
+kind: Infrastructure
+metadata:
+  name: signoz
+spec:
+  deployment:
+    platform: aws
+    mode: kubernetes
+    flavor: terraform
+  resource:
+    kind: Installation
+    spec:
+      name: signoz
