+++ docs/examples/aws/kubernetes/terraform/pours/infrastructure/outputs.tf.json
@@ -1,0 +1,41 @@
+{
+  "output": {
+    "cluster_name": {
+      "description": "Name of the EKS cluster",
+      "value": "${aws_eks_cluster.main.name}"
+    },
+    "cluster_endpoint": {
+      "description": "Endpoint for the EKS cluster API server",
+      "value": "${aws_eks_cluster.main.endpoint}"
+    },
+    "cluster_ca_certificate": {
+      "description": "Base64-encoded certificate authority data for the EKS cluster",
+      "value": "${aws_eks_cluster.main.certificate_authority[0].data}",
+      "sensitive": true
+    },
+    "cluster_version": {
+      "description": "Kubernetes version of the EKS cluster",
+      "value": "${aws_eks_cluster.main.version}"
+    },
+    "vpc_id": {
+      "description": "ID of the VPC",
+      "value": "${aws_vpc.main.id}"
+    },
+    "private_subnet_ids": {
+      "description": "IDs of the private subnets",
+      "value": "${aws_subnet.private[*].id}"
+    },
+    "public_subnet_ids": {
+      "description": "IDs of the public subnets",
+      "value": "${aws_subnet.public[*].id}"
+    },
+    "node_group_default_arn": {
+      "description": "ARN of the default node group",
+      "value": "${aws_eks_node_group.default.arn}"
+    },
+    "node_group_default_status": {
+      "description": "Status of the default node group",
+      "value": "${aws_eks_node_group.default.status}"
+    }
+  }
+}
