+++ docs/examples/aws/kubernetes/terraform/pours/infrastructure/providers.tf.json
@@ -1,0 +1,14 @@
+{
+  "terraform": {
+    "required_version": ">= 1.0.0",
+    "required_providers": {
+      "aws": {
+        "source": "hashicorp/aws",
+        "version": "~> 5.0"
+      }
+    }
+  },
+  "provider": {
+    "aws": [{}]
+  }
+}
